RAA001 рент апратмент готов

+++ function_module/src/main/java/com/first/function_module/repository/dao/ApartmentDao.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface ApartmentDao {
 
-    List<ApartmentEntity> findApartmentByCountOfPeople(Double countOfPeople);
+    List<ApartmentEntity> findApartmentByCountOfPeople(Integer countOfPeople);
 }
